Refactor: Split Unit of Work implementation into separate modules for better structure and modularity
